Slight adjustments to Graph related APIs. Add some unit testing for EdgeWeightedGraph.

+++ src/prj/clark/cs/dsa/algo/graph/Edge.java
@@ -5,7 +5,7 @@ public class Edge implements Comparable<Edge> {
     private final int w;
     private final double weight;
 
-    public Edge(int v, int w, int weight) {
+    public Edge(int v, int w, double weight) {
         this.v = v;
         this.w = w;
         this.weight = weight;
